LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/alignment/flex-align-baseline-line-clamp-002.tentative-expected.txt




































FAIL .target > * 1 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 2
PASS .target > * 3
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="65"><span></span></div>
offsetLeft expected 65 but got 40
PASS .target > * 6
PASS .target > * 7
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 14
PASS .target > * 15
PASS .target > * 16
FAIL .target > * 17 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 18
PASS .target > * 19
PASS .target > * 20
FAIL .target > * 21 assert_equals:
<div data-offset-x="105"><span></span></div>
offsetLeft expected 105 but got 40
PASS .target > * 22
PASS .target > * 23
PASS .target > * 24
